add settings to toggle native emoji size
We don't need two versions of the same plugin. I added a settings menu to toggle using native size or not.

I had also added a number input to restrict it to a custom size but realized the links broke unless the size was certain numbers, and people probably wouldn't want to set it to anything other than either native size or the default 48px used by Discord anyway,